requirements.txt
allure-pytest==2.14.2
allure-python-commons==2.14.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.5.7
colorama==0.4.6
lxml==4.9.3
more-itertools==9.1.0
namedlist==1.8
pluggy==1.2.0
pytest==8.3.5
requests==2.31.0
six==1.16.0
xlrd==2.0.1
